tests/kms_writeback: Iterate over only non-null pipes
Currently, kms_writeback_get_output iterates over all entries of
display->pipes, regardless of if the pipe is valid. This could lead to a
null dereference in igt_output_set_pipe() as the loop does not check for
null pipes. To avoid this issue, use the for_each_pipe method to iterate
over only non-null pipes.

Signed-off-by: Rohith Iyer <quic_rohiiyer@quicinc.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>