lib/igt_kms: Add a helper function to check Bigjoiner support
Create a helper function to check that the system supports the
given crtc/connector mode(s).

Example:
* Pipe-D won't support Bigjoiner, hence we can't use the connector
  modes greater than 5K on Pipe-D
* To use 8K mode on a pipe, then the consecutive pipe must be free.

The Kernel is expected to reject the invalid combo (which must be
validated as a scenario separately). So, this helper function checks
the validity of the combo to avoid failures.

To use this helper, each individual subtest needs to set the @pipe
to a specific @output by igt_output_set_pipe() and call this helper
to check the validity of the combo.

V2:
* Rebase
* Add support to handle fused pipes
V3:
* Reject the bigjoiner if the consecutive pipe is not available

Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>