tests/kms_sysfs_edid_timing: clarify test description and thresholds
The description of this test was cryptic, mentioning a very old patch
and so on.  Clarify what the test actually does.

While at it, improve the assertion messages and clarify the threshold
macro names.

Additionally, change the mean time threshold per connector to
something shorter than the threshold per probe, since the former
cannot happen without the latter already causing an assertion failure.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-by: Mika Kahola <mika.kahola@intel.com>