tools/intel_vbt_decode: update vbt defs from kernel
Sync up the the VBT definitions from kernel commit
ac7af83ae08b ("drm/i915: Rename some VBT bits")
and adjust the actual code to match.

Note that 'slave_port' is gone now, so we'll just throw it out,
and the DP redriver stuff is now defined as bitfields which need
to be properly decoded.

Acked-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>