tools/intel_vbt_decode: Clean up driver features block decoding
Print all the stuff from the driver features block, and print
it in the same order as the spec lists it to make it a bit
easier to compare the two.

Also use consistent naming for a bunch of the fields that
exist in both the driver features block and the power conservation
block.

Acked-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>