i915/i915_module_load: Randomise fault injection
Reduce the number of faults we inject during module reload to bring the
test runtime back to a reasonable level, and rely on multiple CI runs to
establish the full picture. This reduces the efficacy of testing (i.e.
premerge will probably not prevent inclusion of a buggy patch), but
error handling on module reload is not a critical factor for driver
health -- a broken patch that explodes in an error path is unlikely to
prevent CI from detecting critical issues, and with multiple runs we
should still be able to detect and fix issues before merging.

Fixes: https://gitlab.freedesktop.org/drm/intel/-/issues/6393

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>