i915/i915_module_load: Raise a fatal-error on failing to unload a module
If we fail to unload the module after the fault-injection test, we leave
behind a dangerous modparam that will affect all subsequent tests.
Better to declare the run void rather than have to triage all the
bizarre bugs.

Sometimes the fault injection module load test timeouts out, causing
igt_runner to cancel the test with a SIGQUIT. This unfortunately causes
us to abort in the middle of the test leaving the module in a broken
state. That broken module is then used for subsequent tests, causing a
cascade of failures.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Jonathan Cavitt <jonathan.cavitt@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>