lib: Force rescan when module is loaded
Whenever we load a drm module we need to force a rescan of igt devices
so we update the list. Otherwise any place trying to open the device and
then falling back to loading the driver and loading it, will not have a
proper fallback.  Currently this makes igt@i915_pm_rpm@module-reload
to skip on DG2 as the previous test makes igt@i915_pm_rpm to cause a
module reload.

Fixes: https://gitlab.freedesktop.org/drm/intel/-/issues/5174
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Swati Sharma <swati2.sharma@intel.com>