tests/i915/perf_pmu: Avoid unordered sampling with SEMA_WAIT
On Broadwell/Braswell, we sometimes see that the MI_SEMAPHORE_WAIT reads
from memory before the MI_STORE_DWORD_IMM was written.  This causes the
semaphore to complete too early, and the test asserts as the batch is
completed before the measurement is begun.

We can use more MI_STORE_DWORD_IMM before the MI_SEMAPHORE_WAIT to ensure
the MI pipelined is flushed and the write to memory is visible before
our first read by the semaphore-wait. However, we can also just change
our MI_SEMAPHORE_WAIT operation to not break if it sees the initial 0 by
waiting for the value to change to 2. (Now the value starts at 0, is set
to 1 by the batch to indicate it has started and the caller should start
its sampling, and then set to 2 by the caller as it finishes sampling
the semaphore busyness.)

Closes: https://gitlab.freedesktop.org/drm/intel/issues/2383
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>