tests: Avoid the name "all"
If using a subtest or a dynamic subtest name "all", the piglit format
name will contain the string "@all". When filing bugs for test
failures, one needs extra care not to accidentally ping every
developer on gitlab.

Along with having that extra care, try to avoid the possibility of
accidents by renaming everything with just "all" in their name to a
more descriptive form that explains what the list of "all" contains.

Closes: https://gitlab.freedesktop.org/drm/igt-gpu-tools/-/issues/121
Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>