i915/i915_pm_rps: Write values in pm_rps_exit_handler only for min-max configs
pm_rps_exit_handler writes min and max frequencies to the corresponding
sysfs files. In a situation where each test initializes only specific
handlers, it's possible that the exit handler will dereference a NULL
value in MIN and MAX file handlers.

To prevent this, check if the exit handler is called by a subtest
that initialized handlers to files describing frequency constraints.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Karolina Drobnik <karolina.drobnik@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>