tests/gem_exec_fence: Restore pre-hang checks in *await-hang scenarios
Commit c8f6aaf32d83 "tests/gem_exec_fence: Check stored values only for
valid workloads" resolved an issue, observed in *await-hang scenarios,
where a fence exposed by an invalid spin batch was signaled asynchronously
to pending checks for depended test batches still waiting for that fence.
Those checks have been disabled, weakening those scenarios.

This change restores the pre-hang checks to the extent possible when the
invalid spin batch may trigger an immediate reset.  If we are lucky enough
to take a snapshot of the object supposed to be still not modified by
store batches after we confirm that the spin batch has started and before
the fence is signaled, we use that copy to verify if the fence dependent
batches are still blocked.  Running the *await-hang subtests multiple
times in CI should build our confidence in their results.

v2: preserve checking the pipeline runs ahead of the hang (Chris)
v3: use a more simple 'best effort' approach suggested by Chris

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mauro Carvalho Chehab <mchehab@kernel.org>