lib/chamelium: Use MST Path property to reprobe an MST connector.
[Why]
On replugs, MST connectors can change their names and IDs. The only
constant thing is their path property.

[How]
If a connector has a PATH prop, it's a MST connector.
On reprobe, look for MST connector with the same PATH prop.

Tested on Volteer with Chamelium V3 connected via a 2-DP-port CableMatters
hub.
Test: kms_chamelium --run-subtest dp-hpd

Signed-off-by: Mark Yacoub <markyacoub@chromium.org>
Reviewed-by: Lyude Paul <lyude@redhat.com>