tests/i915/kms_big_fb: First async flip discarded on i915
Display 13 onwards, i915 uses the first async flip to update the
watermarks as per the watermark optimization. Hence the actual
async flip will happen from the subsequent flips.

To handle this, a dummy async flip has to be done in the async flip
subtests to allow the KMD to perform the watermark related updates
before writing to the surface base address.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>