lib/igt_kmod: remove the conditional audio removal code
It was expected that kernel/modules would contain a fix for
lsmod to properly report bind dependencies between snd_hda_intel
and i915, but, unfortunately:

- This didn't happen yet;
- Kernel 5.20 won't exist.

This is now causing issues, as drm-tip is now based on 6.0-rc1.

As IGT is working fine without that, let's remove the code.
This patch can be reverted once upstream gets fixed.

Reported-by: Andi Shyti <andi.shyti@linux.intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>
Tested-by: Andi Shyti <andi.shyti@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>