igt: Allow overriding the commit hash in the version string
By default, the latest Git commit hash is always used to create the
version string embedded in libigt - if libigt changes, most targets are
marked as "dirty" and need to be relinked. This makes it difficult to
easily compare two IGT builds on binaries alone.

This commit introduces a new Meson option, -Dversion_hash="your text!",
which allows overriding this hash, for instance under CI builds.

Signed-off-by: Ryszard Knop <ryszard.knop@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>