scripts/code_cov_parse_info: print common coverage and extra per function
While analyzing the data, it is nice to have a quick summary about
how much each set of the analyzed data contributes to the total.

So, generate a function summary report with the interception of
all the sets ("common function coverage"), plus the difference
that an individual test group contributes.

This is useful to provide a quick feedback on several situations:

- When comparing FULL with BAT:
	- how many functions are added on FULL?
	- are there functions that are only on BAT?

- When comparing different machines:
	- How many functions each individual machine has over
	  the common group?

- When comparing a group of tests:
	- How many functions an individual test adds for the
	  coverage that weren't covered at the common group yet?

Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>