Chamelium: Handle getting port connector if it's MST.
[Why]
When getting a connector for a port, the connector ID might be invalid
after a replug if it's an MST connector. Getting the connector with just
the connector ID will not always work.

[How]
Check if the connector is MST by checking the port's connector path.
If it is, iterate through all connectors until a connector with a
matching path is found.

TEST=./kms_chamelium --run-subtest dp-hpd [on intel Volteer board with
Chamelium V3 connected]

Signed-off-by: Mark Yacoub <markyacoub@chromium.org>
Reviewed-by: Lyude Paul <lyude@redhat.com>