tests/kms_content_protection: Fix MST subtests
Add a modeset before checking output is HDCP capable.
Test currently skips on multidisplay MST config telling no HDCP capable
output is found because we're checking if output is HDCP capable
before doing a modeset on that output.

Try other pipe/output combo for MST if failure is seen on one, as the
test is not dependent on the combination as long as each output has a
valid pipe.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>