tests/kms_content_protection: Test Cleanup
Reset display to clean the state before starting the subtest.

Move the clean up part to a separate function so that cleanup code is
run even if there is a failure in the subtest. This prepares the test
for conversion to dynamic subtests.

Remove redundant mode override as we're anyway overriding with the
default mode.

Check if platform is capable of atomic commit before choosing
commit style in cleanup function.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>