Replace instances of mmap64() with mmap()
The mmap64() function exists to point 32-bit architectures to an mmap
interface with the capability to address a 64-bit memory address space.
As 32-bit machines are replaced by 64-bit, there is no longer a need to
keep mmap64() around. This is especially apparent in operating systems
like FreeBSD that exclude the mmap64() function altogether. This patch
replaces all instances of mmap64() with mmap().

Signed-off-by: Jake Freeland <jfree@freebsd.org>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>