tests/kms_busy: Ensure GPU reset when waiting for a new FB during modeset
When waiting for a modeset's new framebuffer the timeout for this FB's
fence to signal is determined by the kernel's
CONFIG_DRM_I915_FENCE_TIMEOUT which is 10 sec by default (and in the
current CI kconfig). The GPU reset timeout in turn is determined by
CONFIG_DRM_I915_HEARTBEAT_INTERVAL which is 2.5 sec by default (and in
the current CI kconfig). By default a spinner batch - like the one
created by test_atomic_commit_hang() - will cause a GPU reset that is 5
times the above heartbeat interval (the interval rounded to seconds
first, that is 15 seconds atm).

Based on the above during extended-modeset-hang-newfb-with-reset the
enabling modeset part for instance will cause a fence timeout before the
presumed GPU hang would be detected, so the corresponding commit will
complete without the expected GPU reset.

Note that extended-modeset-hang-oldfb-with-reset for instance is not
affected by this problem, since waiting for the old framebuffer's fence
to signal happens without a fence timeout (timing out instead at the
GPU hang detect/reset).

Fix the above by reducing the heartbeat interval to 250 msec,
corresponding to a ~1.3 sec GPU hang detection timeout. This also speeds
up the GPU hang subtests.

Testcase: igt@kms_busy@extended-modeset-hang-newfb-with-reset
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>