tests/i915/gem_ctx_engines: Avoid SET_CACHING ioctl during fork
Some subtests for gem_ctx_engines require the SET_CACHING
ioctl to run.  However, the function that calls it is
wrapped by an igt_require and thus causes a skip when the
ioctl fails.  Unfortunately, MTL currently fails this check
with -EOPNOTSUPP.

While the ioctl failing and resulting in a skip wouldn't
normally be a problem, the failing call is occasionally
made within a forked child process, which causes an error.
Fortunately, use of the ioctl can be avoided by using wc
instead of cpu for mmapping and domain operations.

v2: Use gem_mmap__device_coherent instead of gem_mmap__wc

Cc: Fei Yang <fei.yang@intel.com>
Cc: Chris Wilson <chris.p.wilson@intel.com>
Signed-off-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Acked-by: Priyanka Dandamudi <priyanka.dandamudi@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>