tests/vc4: Skip VC4 tests if they are running on BCM2711/RaspberryPi4
Currently, if the VC4 tests are run on BCM2711/Raspberry Pi 4, they will
fail with no warning. So, add igt_require to the VC4 tests to check if
VC4 has rendering capabilities before running the tests.

In order to check if VC4 is running on BCM2711/Raspberry Pi 4, create a
function that checks if the vc4 driver has syncobj capabilities. If not so,
it means that the tests are being run on BCM2711/Raspberry Pi 4, as vc5
doesn't have syncobj capabilities.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>
Reviewed-by: Melissa Wen <mwen@igalia.com>