runner: Don't read journal at all if socket comms are used
When socket communications are used, the journal is left empty. When
processing results, reading an empty journal will add a test result
for the requested test, which can be incorrect when using testlists
with wildcards.

Testlists with wildcards were not the intended use of testlists but
they are used and thus this is considered a fix for a regression.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Nicholas Choi <Nicholas.Choi@amd.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>