tools/intel_gtt: stop confusing GTT and BAR sizes
The GTT iteration loop would erroneously check "< gtt_size" as an
indicator for when it should stop, which would make it exit the loop
right after dumping 2-8 mib of address space depending on the
GPU generation.

Calculate the actual size of GTT in 'gtt_max_addr', and rewrite the
iteration loop to use it. While at it, also make virtual addresses
64-bit so that they don't wrap around at the 4GiB mark.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>