lib/intel_batchbuffer: Use correct engine id in exec_blit()
exec_blit() assumes fixed id of the blitter engine. This is
not correct as the userspace contexts might have engines
stored in a different order.

Make exec_blit accept context configuration that describes
the engine layout, and use it to find the proper engine id.
Update signatures of igt_blitter_src_copy() and
igt_blitter_fast_copy__raw() to reflect that change. Correct
function calls in the tests. Move find_engine() definition
higher up, so it can be used in exec_blit().

Signed-off-by: Karolina Stolarek <karolina.stolarek@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>