lib/v3d: Add V3D packet description
The pack header is a generated file that contains packets, enums, and
structures for a platform. In order to make a valid job submission to
V3D, some data structures are needed. Therefore, add V3D's packet header
for IGT use.

This pack header was generated with v3d_packet_v33.xml through a Python
script available at Mesa [1]. It was generated for V3D version 4.2.

[1] https://gitlab.freedesktop.org/mesa/mesa/-/blob/22.3/src/broadcom/cle/gen_pack_header.py

Reviewed-by: Melissa Wen <mwen@igalia.com>
Signed-off-by: Maíra Canal <mcanal@igalia.com>