tests/i915/gem_basic: add multigpu-create-close subtest
Add one subtest to run on more than one gpu simultanosly. For
this to work test should be run with --device option on machine
with two or more dgfx cards, for example with:

--device=pci:vendor=Intel,device=discrete,card=all

Test will skip if only one card is present.

v2: remove code for setting gpu_count=1 (Mauro)

Cc: Anna Karas <anna.karas@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Mauro Carvalho Chehab <mauro.chehab@linux.intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab@kernel.org>