lib/igt_sysfs: Generic verification of writable sysfs attributes
Attempt to verify writable sysfs attributes in a generic way, without going
into the specifics of any attribute. The attribute is first written to and
then read back and it is verified that the read value matches the written
value to a tolerance. However, when we try to do this we run into the issue
that a sysfs attribute might have a behavior where the read value is
different from the written value for any reason. For example, attributes
such as power, voltage, frequency and time typically have a linear region
outside which they are clamped (the values saturate). Therefore for such
attributes read values match the written value only in the linear region
and when writing we don't know if we are writing to the linear or to the
clamped region.

Therefore the verification implemented here takes the approach of sweeping
across the range of possible values of the attribute (this is done using
'doubling' rather than linearly) and seeing where there are matches. There
should be at least one match for the verification to have succeeded.

Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Riana Tauro <riana.tauro@intel.com>