tests/v3d_mmap: Improve tests from V3D's Mmap BO IOCTL
The current tests for the V3D's Mmap IOCTL don't cover all the
possible invalid parameters on drm_v3d_mmap_bo. Therefore, add a
subtest to make sure that flags on drm_v3d_mmap_bo is zero and also
add a subtest to test the read/write coherency of a newly mapped bo.
Moreover, also adds a check to assure that the mmap's offset is a
multiple of the page size as returned by sysconf(_SC_PAGE_SIZE).

Reviewed-by: Melissa Wen <mwen@igalia.com>
Signed-off-by: Maíra Canal <mcanal@igalia.com>