tests/kms_cursor_crc: Skip for invalid pipe connector connection
Added a skip condition that checks for the valid pipe connection. This is a
generic skip on pipe which checks for valid connector connection mask bit.
This condition helps to skip tests for MSO panels on non supported pipes
(C and D). This condition was missed for cursor-size-change,
cursor-alpha-opaque and cursor-alpha-transparent testcases.

--v2
- Change commit messege  from "Add Gaurd for MSO eDP for Pipe C and D"
to "Skip condition for non valid pipe connector connection" [Kamil]
- skip condition using igt_pipe_connector_valid() [Juha-Pekka]
- fixed tabs [Juha-Pekka]

Signed-off-by: Dnyaneshwar Bhadane <dnyaneshwar.bhadane@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>